+++ src/dcos_e2e/node.py
@@ -684,3 +684,60 @@ class Node:
             transport=transport,
             sudo=sudo,
         )
+
+    def download_file(
+        self,
+        remote_path: Path,
+        local_path: Path,
+        transport: Optional[Transport] = None,
+    ) -> None:
+        """
+        Download a file from this node.
+
+        Args:
+            remote_path: The path on the node to download the file from.
+            local_path: The path on the host to download the file to.
+            transport: The transport to use for communicating with nodes. If
+                ``None``, the ``Node``'s ``default_transport`` is used.
+
+        Raises:
+            ValueError: The ``remote_path`` does not exist. The ``local_path``
+                is an existing file.
+        """
+        transport = transport or self.default_transport
+        user = self.default_user
+        transport = self.default_transport
+        try:
+            self.run(
+                args=['test', '-e', str(remote_path)],
+                user=user,
+                transport=transport,
+                sudo=False,
+            )
+        except subprocess.CalledProcessError:
+            message = (
+                'Failed to download file from remote location "{location}". '
+                'File does not exist.'
+            ).format(location=remote_path)
+            raise ValueError(message)
+
+        if local_path.exists() and local_path.is_file():
+            message = (
+                'Failed to download a file to "{file}". '
+                'A file already exists in that location.'
+            ).format(file=local_path)
+            raise ValueError(message)
+
+        if local_path.exists() and local_path.is_dir():
+            download_file_path = local_path / remote_path.name
+        else:
+            download_file_path = local_path
+
+        node_transport = self._get_node_transport(transport=transport)
+        node_transport.download_file(
+            remote_path=remote_path,
+            local_path=download_file_path,
+            user=user,
+            ssh_key_path=self._ssh_key_path,
+            public_ip_address=self.public_ip_address,
+        )